# src/main/resources/static/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Order Picker API",
    "description": "    This api is made to improve programming skills with the Java language, Spring Boot framework and MySQL Database, in addition\n    the api has unit test with Junit and Mockito.\n    To begin you must create a user with Role USER or EMPLOYEE and run the different endpoints where you can create products, orders and\n    make deliveries according to the role that chose. \n",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/order-picker/orders/{id}": {
      "get": {
        "tags": [
          "order-controller"
        ],
        "summary": "Get an Order by id",
        "operationId": "getOneById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Order ID to search",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order Found",
            "headers": {
              "Authorization": {
                "description": "Token authorization",
                "style": "simple"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Orders"
                }
              }
            }
          },
          "403": {
            "description": "No permission to apply"
          },
          "404": {
            "description": "Order Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Orders": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string"
          },
          "userAddress": {
            "type": "string"
          },
          "userPhone": {
            "type": "string"
          },
          "userEmail": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "orderDescription": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "isDelivered": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}